renamed models package as linearmodel to reduce confusion with generated model
