+++ forecaster.go
@@ -13,7 +13,7 @@ import (
 	"time"
 
 	"github.com/aouyang1/go-forecaster/forecast"
-	"github.com/aouyang1/go-forecaster/models"
+	"github.com/aouyang1/go-forecaster/linearmodel"
 	"github.com/aouyang1/go-forecaster/stats"
 	"github.com/aouyang1/go-forecaster/timedataset"
 	"github.com/go-echarts/go-echarts/v2/components"
@@ -307,17 +307,17 @@ func (f *Forecaster) Predict(t []time.Time) (*Results, error) {
 // Score computes the coefficient of determination of the prediction
 func (f *Forecaster) Score(t []time.Time, y []float64) (float64, error) {
 	if t == nil {
-		return 0.0, fmt.Errorf("no time slice for inference, %w", models.ErrNoDesignMatrix)
+		return 0.0, fmt.Errorf("no time slice for inference, %w", linearmodel.ErrNoDesignMatrix)
 	}
 	if y == nil {
-		return 0.0, fmt.Errorf("no expected values for inference, %w", models.ErrNoTargetMatrix)
+		return 0.0, fmt.Errorf("no expected values for inference, %w", linearmodel.ErrNoTargetMatrix)
 	}
 
 	m := len(t)
 
 	ym := len(y)
 	if m != ym {
-		return 0.0, fmt.Errorf("design matrix has %d rows and target has %d rows, %w", m, ym, models.ErrTargetLenMismatch)
+		return 0.0, fmt.Errorf("design matrix has %d rows and target has %d rows, %w", m, ym, linearmodel.ErrTargetLenMismatch)
 	}
 
 	res, err := f.Predict(t)
