+++ linearmodel/lasso.go
@@ -1,4 +1,4 @@
-package models
+package linearmodel
 
 import (
 	"errors"
