+++ linearmodel/lasso_test.go
@@ -1,4 +1,4 @@
-package models
+package linearmodel
 
 import (
 	"testing"
